基本功能完善

+++ app/src/main/java/com/lutong/activity/Adapter/TaCallBack.java
@@ -1,6 +1,6 @@
 package com.lutong.activity.Adapter;
 
 public interface TaCallBack {
-    void call(int posion, Double data);
+    void call(int posion, Integer data);
     void callDele(int i);
 }
